Fix castle upgrade shadow

+++ src/fheroes2/maps/maps_tiles.h
@@ -181,7 +181,7 @@ namespace Maps
         void AddonsSort( void );
         void Remove( u32 uniqID );
         void RemoveObjectSprite( void );
-        void UpdateObjectSprite( uint32_t uniqID, uint8_t rawTileset, uint8_t newTileset, uint8_t indexChange );
+        void UpdateObjectSprite( uint32_t uniqID, uint8_t rawTileset, uint8_t newTileset, int indexChange );
         void ReplaceObjectSprite( uint32_t uniqID, uint8_t rawTileset, uint8_t newTileset, uint8_t indexToReplace, uint8_t newIndex );
 
         std::string String( void ) const;
